# backend_go/docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  models.History:
    properties:
      date:
        type: string
      id:
        type: integer
      nb_jetons:
        type: integer
      stand_name:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Kermesse:
    properties:
      id:
        type: integer
      name:
        type: string
      organisateurs:
        description: 'Relations Many-to-Many : Organisateurs et participants de la
          kermesse'
        items:
          $ref: '#/definitions/models.User'
        type: array
      participants:
        items:
          $ref: '#/definitions/models.User'
        type: array
      picture:
        type: string
      stands:
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      user_id:
        description: 'Relation Many-to-One : L''utilisateur qui crée la kermesse'
        type: integer
    type: object
  models.Product:
    properties:
      id:
        type: integer
      jetons_requis:
        type: integer
      name:
        type: string
      nb_products:
        type: integer
      picture:
        type: string
      stand_id:
        description: Clé étrangère vers le stand
        type: integer
      type:
        type: string
    type: object
  models.Stand:
    properties:
      conso:
        type: integer
      id:
        type: integer
      jetons_requis:
        type: integer
      kermesses:
        items:
          $ref: '#/definitions/models.Kermesse'
        type: array
      name:
        type: string
      pts_donnees:
        type: integer
      stocks:
        description: Clé étrangère vers Product
        items:
          $ref: '#/definitions/models.Product'
        type: array
      type:
        type: string
      user_id:
        type: integer
    type: object
  models.Transaction:
    properties:
      Quantity:
        type: integer
      date_transaction:
        type: string
      id:
        type: integer
      price:
        type: number
      type:
        type: string
      user_id:
        description: Relations avec l'utilisateur
        type: integer
    type: object
  models.User:
    properties:
      email:
        type: string
      enfants:
        items:
          $ref: '#/definitions/models.User'
        type: array
      firstname:
        type: string
      historique:
        items:
          $ref: '#/definitions/models.History'
        type: array
      id:
        type: integer
      jetons:
        type: integer
      kermesses:
        description: Relations One-to-Many pour les Kermesses et Stands gérés par
          l'utilisateur
        items:
          $ref: '#/definitions/models.Kermesse'
        type: array
      lastname:
        type: string
      parents:
        description: Relations Many-to-Many pour Parents/Enfants
        items:
          $ref: '#/definitions/models.User'
        type: array
      password:
        type: string
      picture:
        type: string
      pts_attribues:
        type: integer
      role:
        description: 1 = ADMIN / 2 = ORGANISATEUR / 3 = TENEUR DE STAND / 4 = PARENT
          / 5 ELEVE
        type: integer
      stands:
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
    type: object
...
  requests.GivePointsRequest:
    properties:
      points:
        type: integer
    type: object
...
  requests.QuantityProductRequest:
    properties:
      quantity:
        type: integer
    type: object
...
  requests.StandRequest:
    properties:
      jetons_requis:
        type: integer
      name:
        type: string
      type:
        type: string
    required:
    - jetons_requis
    - name
    - type
    type: object
info:
  contact: {}
paths:
...
  /create-stand:
    post:
      consumes:
      - application/json
      description: Crée un nouveau stand avec les informations fournies
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Stand à créer
        in: body
        name: stand
        required: true
        schema:
          $ref: '#/definitions/requests.StandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Stand'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Crée un nouveau stand
      tags:
      - Stand
...
  /stands:
    get:
      description: Récupère la liste de tous les stands
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Stand'
            type: array
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Récupère tous les produits
      tags:
      - Stand
  /stands/{id}:
    get:
      description: Récupère les informations d'un utilisateur spécifique
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stand'
        "404":
          description: Stand non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Récupère un stand par ID
      tags:
      - Stand
  /stands/{id}/delete:
    delete:
      description: Supprime un stand spécifique
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de du stand
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: stand non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Supprime un stand par ID
      tags:
      - Stand
  /stands/{id}/interact:
    post:
      consumes:
      - application/json
      description: Permet d'interagir avec un stand en échange de jetons
      parameters:
      - description: ID du stand
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stand'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Interagir avec un stand
      tags:
      - Stand
  /stands/{id}/update:
    put:
      consumes:
      - application/json
      description: Met à jour les informations d'un stand spécifique
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      - description: Stand à mettre à jours
        in: body
        name: stand
        required: true
        schema:
          $ref: '#/definitions/models.Stand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stand'
        "404":
          description: Stand non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Met à jour un stand par ID
      tags:
      - Stand
  /stands/{stand_id}/products/{product_id}/buy:
    post:
      consumes:
      - application/json
      description: Permet à un utilisateur d'acheter un produit sur un stand avec
        ses jetons
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID du stand
        in: path
        name: stand_id
        required: true
        type: integer
      - description: ID du produit
        in: path
        name: product_id
        required: true
        type: integer
      - description: Quantité de produit à acheter
        in: body
        name: quantity
        required: true
        schema:
          $ref: '#/definitions/requests.QuantityProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      summary: Achat d'un produit sur un stand
      tags:
      - Stand
  /stands/{stand_id}/users/{user_id}/points:
    post:
      consumes:
      - application/json
      description: Un stand peut attribuer des points à un utilisateur en fonction
        du type de stand
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID du stand
        in: path
        name: stand_id
        required: true
        type: integer
      - description: ID de l'utilisateur
        in: path
        name: user_id
        required: true
        type: integer
      - description: Nombre de points à attribuer
        in: body
        name: points
        required: true
        schema:
          $ref: '#/definitions/requests.GivePointsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      summary: Attribue des points à un utilisateur depuis un stand
      tags:
      - Stand
...
swagger: "2.0"